MLOps/model-deployment/pyproject.toml
[project]
name = "ml-dl-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "transformers>=4.40",
    "torch>=2.2",
    "tensorflow>=2.16",
    "pillow>=10.0",
    "numpy>=1.26",
    "python-multipart>=0.0.9",
    "streamlit>=1.35",
    "streamlit-drawable-canvas>=0.9",
    "requests>=2.31",
]

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.27"]
optional = ["watchdog>=4.0"]
